moves location of adapters  - more convenient when next to class (or even inside a class ?)
