+++ src/main/java/sc/fiji/bdvpg/bdv/supplier/IBdvSupplier.java
@@ -2,7 +2,6 @@ package sc.fiji.bdvpg.bdv.supplier;
 
 import bdv.util.BdvHandle;
 import org.scijava.plugin.SciJavaPlugin;
-import sc.fiji.bdvpg.scijava.adapter.bdv.DefaultBdvSupplierAdapter;
 
 import java.util.function.Supplier;
